rename: 🚚 (api) move provider schema files
