+++ apps/api/src/domain/oauth-account/provider/index.ts
@@ -1,0 +1,2 @@
+export type { OAuthProvider } from "./provider.type";
+export { oAuthProviderSchema } from "./provider.schema";
