+++ apps/api/src/domain/oauth-account/provider/provider.type.ts
@@ -1,4 +1,4 @@
 import type { Static } from "elysia";
-import type { oAuthProviderSchema } from "../schema/provider.schema";
+import type { oAuthProviderSchema } from "./provider.schema";
 
 export type OAuthProvider = Static<typeof oAuthProviderSchema>;
